+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway/Cert.hs
@@ -145,8 +145,8 @@ instance SpecTranslate ConwayEra (EpochState ConwayEra) where
       ratifyState = getRatifyState $ utxosGovState lsUTxOState
       govActions = toList $ lsUTxOState ^. utxosGovStateL . proposalsGovStateL . pPropsL
 
-instance SpecTranslate ConwayEra SnapShots where
-  type SpecRep ConwayEra SnapShots = Agda.Snapshots
+instance SpecTranslate ConwayEra (SnapShots era) where
+  type SpecRep ConwayEra (SnapShots era) = Agda.Snapshots
 
   toSpecRep (SnapShots {..}) =
     Agda.MkSnapshots