+++ eras/allegra/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.10.0.0
 
+* Add `TranslateEra` instance for `SnapShots`
 * Rename rule types and deprecate the old names:
   - `AllegraUTXO` -> `UTXO`
   - `AllegraUTXOW` -> `UTXOW`